feat(termination) - finish termination reasons diaslog
