+++ src/flows/Termination/index.ts
@@ -1,3 +1,4 @@
 export { TerminationFlow } from './TerminationFlow';
+export { TerminationReasonsDetailContent } from './TerminationReasonsDetailContent/TerminationReasonsDetailContent';
 export type { TerminationRenderProps } from './TerminationFlow';
 export type { TerminationFormValues } from './types';
